image_processing/requirements.txt
numpy
tensorflow-gpu
dlib
opencv-contrib-python
flask
imutils
socketIO-client